refactor(rust): rename decryptor worker to decryptor handler
