+++ implementations/rust/ockam/ockam_identity/src/secure_channel/api.rs
@@ -16,11 +16,11 @@ pub enum EncryptionResponse {
     Err(Error),
 }
 
-/// Request type for `DecryptorWorker` API Address
+/// Request type for `Decryptor` API Address (the `Decryptor` is accessible through the `HandshakeWorker`)
 #[derive(Serialize, Deserialize, Message)]
 pub struct DecryptionRequest(pub Vec<u8>);
 
-/// Response type for `DecryptorWorker` API Address
+/// Response type for `Decryptor` API Address (the `Decryptor` is accessible through the `HandshakeWorker`)
 #[derive(Serialize, Deserialize, Message)]
 pub enum DecryptionResponse {
     /// Success
