+++ implementations/rust/ockam/ockam_identity/src/secure_channel/decryptor.rs
@@ -1,12 +1,121 @@
-use crate::identity::IdentityError;
 use crate::secure_channel::encryptor::{Encryptor, KEY_RENEWAL_INTERVAL};
 use crate::secure_channel::nonce_tracker::NonceTracker;
+use crate::secure_channel::Addresses;
 use crate::XXInitializedVault;
+use crate::{
+    DecryptionRequest, DecryptionResponse, IdentityError, IdentityIdentifier,
+    IdentitySecureChannelLocalInfo,
+};
 use ockam_core::compat::sync::Arc;
 use ockam_core::compat::vec::Vec;
-use ockam_core::Result;
+use ockam_core::{Any, Result, Routed, TransportMessage};
+use ockam_core::{Decodable, LocalMessage};
+use ockam_node::Context;
 use ockam_vault::KeyId;
+use tracing::debug;
 use tracing::warn;
+
+pub(crate) struct DecryptorHandler {
+    //for debug purposes only
+    pub(crate) role: &'static str,
+    pub(crate) addresses: Addresses,
+    pub(crate) their_identity_id: IdentityIdentifier,
+    pub(crate) decryptor: Decryptor,
+}
+
+impl DecryptorHandler {
+    pub fn new(
+        role: &'static str,
+        addresses: Addresses,
+        key: KeyId,
+        vault: Arc<dyn XXInitializedVault>,
+        their_identity_id: IdentityIdentifier,
+    ) -> Self {
+        Self {
+            role,
+            addresses,
+            their_identity_id,
+            decryptor: Decryptor::new(key, vault),
+        }
+    }
+
+    pub(crate) async fn handle_decrypt_api(
+        &mut self,
+        ctx: &mut Context,
+        msg: Routed<Any>,
+    ) -> Result<()> {
+        debug!(
+            "SecureChannel {} received Decrypt API {}",
+            self.role, &self.addresses.decryptor_remote
+        );
+
+        let return_route = msg.return_route();
+
+        // Decode raw payload binary
+        let request = DecryptionRequest::decode(&msg.into_transport_message().payload)?;
+
+        // Decrypt the binary
+        let decrypted_payload = self.decryptor.decrypt(&request.0).await;
+
+        let response = match decrypted_payload {
+            Ok(payload) => DecryptionResponse::Ok(payload),
+            Err(err) => DecryptionResponse::Err(err),
+        };
+
+        // Send reply to the caller
+        ctx.send_from_address(return_route, response, self.addresses.decryptor_api.clone())
+            .await?;
+
+        Ok(())
+    }
+
+    pub(crate) async fn handle_decrypt(
+        &mut self,
+        ctx: &mut Context,
+        msg: Routed<Any>,
+    ) -> Result<()> {
+        debug!(
+            "SecureChannel {} received Decrypt {}",
+            self.role, &self.addresses.decryptor_remote
+        );
+
+        // Decode raw payload binary
+        let payload = Vec::<u8>::decode(&msg.into_transport_message().payload)?;
+
+        // Decrypt the binary
+        let decrypted_payload = self.decryptor.decrypt(&payload).await?;
+
+        // Encrypted data should be a TransportMessage
+        let mut transport_message = TransportMessage::decode(&decrypted_payload)?;
+
+        // Add encryptor hop in the return_route (instead of our address)
+        transport_message
+            .return_route
+            .modify()
+            .prepend(self.addresses.encryptor.clone());
+
+        // Mark message LocalInfo with IdentitySecureChannelLocalInfo,
+        // replacing any pre-existing entries
+        let local_info =
+            IdentitySecureChannelLocalInfo::mark(vec![], self.their_identity_id.clone())?;
+
+        let msg = LocalMessage::new(transport_message, local_info);
+
+        match ctx
+            .forward_from_address(msg, self.addresses.decryptor_internal.clone())
+            .await
+        {
+            Ok(_) => Ok(()),
+            Err(err) => {
+                warn!(
+                    "{} forwarding decrypted message from {}",
+                    err, &self.addresses.encryptor
+                );
+                Ok(())
+            }
+        }
+    }
+}
 
 pub(crate) struct Decryptor {
     current_key: KeyId,
@@ -18,6 +127,16 @@ pub(crate) struct Decryptor {
 }
 
 impl Decryptor {
+    pub fn new(key: KeyId, vault: Arc<dyn XXInitializedVault>) -> Self {
+        Self {
+            current_key: key,
+            current_key_nonce: 0,
+            previous_key: None,
+            vault,
+            nonce_tracker: NonceTracker::new(),
+        }
+    }
+
     /// Restore 12-byte nonce needed for AES GCM from 8 byte that we use for noise
     fn convert_nonce_from_small(b: &[u8]) -> Result<(u64, [u8; 12])> {
         let bytes: [u8; 8] = b.try_into().map_err(|_| IdentityError::InvalidNonce)?;
@@ -81,16 +200,6 @@ impl Decryptor {
             }
 
             result
-        }
-    }
-
-    pub fn new(key: KeyId, vault: Arc<dyn XXInitializedVault>) -> Self {
-        Self {
-            current_key: key,
-            current_key_nonce: 0,
-            previous_key: None,
-            vault,
-            nonce_tracker: NonceTracker::new(),
         }
     }
 }
